Separated the scene and sim views so I could start on the glass layers tab
